fix(ui): outside modal click restriction modal update
