--- datahub-web-react/src/app/policy/PolicyOutsideAlerter.tsx
@@ -1,43 +1,0 @@
-import React, { useEffect, useRef } from 'react';
-import { Modal } from 'antd';
-
-interface OutsideAlerterType {
-    children: React.ReactNode;
-    closeModal: () => void;
-}
-
-/**
- * Hook that alerts clicks outside of the passed ref
- */
-export function useOutsideAlerter(ref, closeModal) {
-    useEffect(() => {
-        function handleClickOutside(event) {
-            if (ref.current && event.target?.firstElementChild?.classList.contains('PolicyBuilderModal')) {
-                Modal.confirm({
-                    title: 'Exit Policy Editor',
-                    content: `Are you sure you want to exit policy editor? All changes will be lost`,
-                    onOk() {
-                        closeModal();
-                    },
-                    onCancel() {},
-                    okText: 'Yes',
-                    maskClosable: true,
-                    closable: true,
-                });
-            }
-        }
-        // Bind the event listener
-        document.addEventListener('mousedown', handleClickOutside);
-        return () => {
-            // Unbind the event listener on clean up
-            document.removeEventListener('mousedown', handleClickOutside);
-        };
-    }, [ref, closeModal]);
-}
-
-export default function PolicyOutsideAlerter({ children, closeModal }: OutsideAlerterType) {
-    const wrapperRef = useRef(null);
-    useOutsideAlerter(wrapperRef, closeModal);
-
-    return <div ref={wrapperRef}>{children}</div>;
-}
